reorganized backend directory to try to keep everything well structured.
