+++ backend/game_pieces/levels.py
@@ -1,5 +1,5 @@
 """This file contains metadata describing the swarms in each level."""
-from creep import CREEP_TYPES
+from game_pieces.creep import CREEP_TYPES
 level_one = {'creep_type': CREEP_TYPES['basic'], 'amount': 3}
 level_two = {'creep_type': CREEP_TYPES['basic'], 'amount': 5}
 level_three = {'creep_type': CREEP_TYPES['basic'], 'amount': 7}
