--- backend/server.py
@@ -1,1 +1,0 @@
-# pretend a server with newtork capabilities lives here
